pyproject.toml
[tool.poetry]
name = "hunta-backend"
version = "1.0.0"
description = "Backend for Hunta using FastAPI"
authors = ["JakeHunta"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pillow = "^10.0.0"
requests = "^2.31.0"
torch = "^2.3.0"
scikit-learn = "^1.5.0"
python-dotenv = "^1.0.1"
torchvision = "^0.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
